Added calibration (extrinsic and room)
Intrinsic calibration is still missing, but shouldn't be too hard

Camera positions are calibrated in two phases:
1. Extrinsic: Hold Calibration Marker close to each pair of cameras, so that they can figure out their relation
2. Room: Put the Calibration Marker in the center of the room, so that they find an absolute origin
	(this will later be extended to multiple positions on the floor to make the ground plane more stable)
After 2. concludes, not only the center position of each camera is used to position them (as cameras are potentially far away from the origin),
but instead all possible paths to other cameras are also taken into account and weighted according to the weakest link.
This not only decreases the noise more consistently but also prioritises relations between cameras to be correct rather than just each camera position and the relations being slightly off, since the relations need to be more accurate for tracking.
This works really well in minimizing camera error.

This can be tested in simulation right now, hit 'b' to finalize the current phase and check output and then 'n' to start the next calibration phase.